please xd

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
  */
 void print_binary(unsigned long int n)
 {
-	unsigned int mask;
+	unsigned long int mask;
 
 	if (n == 0)
 	{
